+++ R/aggregate.R
@@ -71,6 +71,7 @@ aggregate.stars = function(x, by, FUN, ..., drop = FALSE, join = st_intersects,
 		as_points = any(st_dimension(by) == 2, na.rm = TRUE), rightmost.closed = FALSE,
 		left.open = FALSE, exact = FALSE) {
 
+	fn_name = substr(deparse1(substitute(FUN)), 1, 20)
 	classes = c("sf", "sfc", "POSIXct", "Date", "PCICt", "character", "function", "stars")
 	if (!is.function(by) && !inherits(by, classes))
 		stop(paste("currently, only `by' arguments of class", 
@@ -92,7 +93,7 @@ aggregate.stars = function(x, by, FUN, ..., drop = FALSE, join = st_intersects,
         	stop("package exactextractr required, please install it first") # nocov
 		x = st_upfront(x)
 		d = st_dimensions(x)[1:2]
-		r = st_as_stars(list(array(1, dim = dim(d))), dimensions = d)
+		r = st_as_stars(list(a = array(1, dim = dim(d))), dimensions = d)
 		e = exactextractr::coverage_fraction(as(r, "Raster"), by)
 		st = do.call(raster::stack, e)
 		m = raster::getValues(st)
@@ -208,8 +209,26 @@ aggregate.stars = function(x, by, FUN, ..., drop = FALSE, join = st_intersects,
 			geom
 	if (drop_y)
 		d = d[-2] # y
+	
+	# suppose FUN resulted in more than one value?
+	if ((r <- prod(dim(x[[1]])) / prod(dim(d))) > 1) {
+		if (r %% 1 != 0)
+			stop("unexpected array size: does FUN return a consistent number of values?")
+		a = attributes(d)
+		values = if (is.null(rn <- rownames(x[[1]])))
+				seq_len(r)
+			else
+				rn[1:r]
+		d = append(d, list(create_dimension(values = values)))
+		n = length(d)
+		names(d)[n] = fn_name
+		d = d[c(n, 1:(n-1))]
+		attr(d, "raster") = a$raster
+		attr(d, "class") = a$class
+		newdim = setNames(c(r, length(by), dims[-(1:ndims)]), names(d))
+	} else
+		newdim = setNames(c(length(by), dims[-(1:ndims)]), names(d))
 
-	newdim = setNames(c(sfc = length(by), dims[-(1:ndims)]), names(d))
 	st_stars(lapply(x, structure, dim = newdim), dimensions = d)
 }