+++ R/plot.R
@@ -6,6 +6,34 @@ make_label = function(x, i = 1) {
 		names(x)[i]
 }
 
+#kw_dflt = function(x, key.pos) {
+#	if (is.null(key.pos) || key.pos <= 0)
+#		lcm(0)
+#	else if (key.pos %in% c(2,4) && is.factor(x[[1]]))
+#		lcm(max(strwidth(levels(x[[1]]), "inches")) * 2.54 * 1.1 + par("ps")/12) # cm
+#	else
+#		lcm(1.8 * par("ps")/12)
+#}
+# copy from sf:
+kw_dflt = function(x, key.pos) {
+	if (is.null(key.pos) || key.pos[1] == 0) # no key:
+		return(lcm(0))
+
+	font_scale = par("ps") / 12
+	if (key.pos[1] == -1)
+		lcm(1.8 * font_scale)
+	else if (key.pos[1] %in% c(2, 4) && (is.character(x[[1]]) || is.factor(x[[1]]))) {
+		strings = if (is.factor(x[[1]]))
+				levels(x[[1]])
+			else
+				x[[1]]
+		lcm(cm(max(strwidth(strings, "inches"))) * 1.3 + font_scale) # cm
+		#lcm(cm(max(strwidth(strings, "inches"))) * 1.3) # cm
+	} else
+		lcm(1.8 * font_scale)
+}
+
+
 #' plot stars object, with subplots for each level of first non-spatial dimension
 #'
 #' plot stars object, with subplots for each level of first non-spatial dimension, and customization of legend key
@@ -18,18 +46,19 @@ make_label = function(x, i = 1) {
 #' @param axes logical; should axes and box be added to the plot?
 #' @param downsample logical or numeric; if \code{TRUE} will try to plot not many more pixels than actually are visible, if \code{FALSE}, no downsampling takes place, if numeric, the number of pixels/lines/bands etc that will be skipped; see Details.
 #' @param nbreaks number of color breaks; should be one more than number of colors. If missing and \code{col} is specified, it is derived from that.
-#' @param breaks actual color breaks, or a method name used for \link[classInt]{classIntervals}.
+#' @param breaks numeric vector with actual color breaks, or a \code{style} name used in \link[classInt]{classIntervals}.
 #' @param col colors to use for grid cells, or color palette function
 #' @param ... further arguments: for \code{plot}, passed on to \code{image.stars}; for \code{image}, passed on to \code{image.default} or \code{rasterImage}.
-#' @param key.pos integer; side to plot a color key: 1 bottom, 2 left, 3 top, 4 right; set to \code{NULL} to omit key. Ignored if multiple columns are plotted in a single function call. Default depends on plot size, map aspect, and, if set, parameter \code{asp}.
+#' @param key.pos numeric; side to plot a color key: 1 bottom, 2 left, 3 top, 4 right; set to \code{NULL} to omit key. Ignored if multiple columns are plotted in a single function call. Default depends on plot size, map aspect, and, if set, parameter \code{asp}. If it has length 2, the second value, ranging from 0 to 1, determines where the key is placed in the available space (default: 0.5, center).
 #' @param key.width amount of space reserved for width of the key (labels); relative or absolute (using lcm)
 #' @param key.length amount of space reserved for length of the key (labels); relative or absolute (using lcm)
 #' @param key.lab character; label for color key in case of multiple subplots, use \code{""} to suppress
 #' @param reset logical; if \code{FALSE}, keep the plot in a mode that allows adding further map elements; if \code{TRUE} restore original mode after plotting
-#' @param box_col color for box around sub-plots; use \code{0} to suppress plotting of boxes around sub-plots.
+#' @param box_col color for box around sub-plots; use \code{NA} to suppress plotting of boxes around sub-plots.
 #' @param center_time logical; if \code{TRUE}, sub-plot titles will show the center of time intervals, otherwise their start
 #' @param hook NULL or function; hook function that will be called on every sub-plot; see examples.
 #' @param mfrow length-2 integer vector with nrows, ncolumns of a composite plot, to override the default layout
+#' @param compact logical; place facets compactly (TRUE), or spread over the plotting device area?
 #' @details
 #' Downsampling: a value for \code{downsample} of 0: no downsampling, 1: after every dimension value (pixel/line/band), one value is skipped (half of the original resolution), 2: after every dimension value, 2 values are skipped (one third of the original resolution), etc. If \code{downsample} is \code{TRUE} or a length 1 numeric vector, downsampling is only applied to the raster [x] and [y] dimensions.
 #'
@@ -51,13 +80,15 @@ make_label = function(x, i = 1) {
 #' }
 #' plot(x, hook = hook2, col = grey(c(.2,.25,.3,.35)))
 #' if (isTRUE(dev.capabilities()$rasterImage == "yes")) {
-#'   lc = system.file("tif/lc.tif", package = "stars")
-#'   plot(read_stars(lc), key.pos=4, key.width=lcm(5))
+#'   lc = read_stars(system.file("tif/lc.tif", package = "stars"))
+#'   levels(lc[[1]]) = abbreviate(levels(lc[[1]]), 6) # so it's not only legend
+#'   plot(lc, key.pos=4)
 #' }
 plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes = FALSE,
 		downsample = TRUE, nbreaks = 11, breaks = "quantile", col = grey(1:(nbreaks-1)/nbreaks),
-		key.pos = get_key_pos(x, ...), key.width = lcm(1.8), key.length = 0.618, key.lab = main,
-		reset = TRUE, box_col = grey(.8), center_time = FALSE, hook = NULL, mfrow = NULL) {
+		key.pos = get_key_pos(x, ...), key.width = kw_dflt(x, key.pos), key.length = 0.618, 
+		key.lab = main, reset = TRUE, box_col = NA, center_time = FALSE, hook = NULL, 
+		mfrow = NULL, compact = TRUE) {
 
 	if (!missing(y))
 		stop("y argument should be missing")
@@ -88,8 +119,9 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 	breaks.missing = missing(breaks)
 	if (missing(nbreaks) && !missing(col))
 		nbreaks = length(col) + 1
-	opar = par()
+	opar = par(no.readonly = TRUE)
 	dots = list(...)
+	need_to_reset = TRUE
 
 	#if (any(dim(x) == 1))
 	#	x = adrop(x)
@@ -142,7 +174,7 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 			if (! isTRUE(dots$add) && ! is.null(key.pos) && !all(is.na(values)) && is.null(dots$rgb) &&
 					(is.factor(values) || length(unique(na.omit(values))) > 1) &&
 					length(col) > 1 && !is_curvilinear(x)) { # plot key?
-				switch(key.pos,
+				switch(key.pos[1],
 					layout(matrix(c(2,1), nrow = 2, ncol = 1), widths = 1, heights = c(1, key.width)),  # 1 bottom
 					layout(matrix(c(1,2), nrow = 1, ncol = 2), widths = c(key.width, 1), heights = 1),  # 2 left
 					layout(matrix(c(1,2), nrow = 2, ncol = 1), widths = 1, heights = c(key.width, 1)),  # 3 top
@@ -156,12 +188,13 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 				else
 					.image_scale(values, col, breaks = breaks, key.pos = key.pos,
 						key.width = key.width, key.length = key.length, axes = axes, ..., lab = key.lab)
-			}
+			} else
+				need_to_reset = FALSE
 
 			# map panel:
-			mar = c(axes * 2.1, axes * 2.1, 1 * !is.null(main), 0)
-			if (!is.null(key.pos) && key.pos %in% 1:4)
-				mar[key.pos] = mar[key.pos] + .5
+			mar = c(axes * 2.1, axes * 2.1, 1.2 * !is.null(main), 0)
+			if (!is.null(key.pos) && key.pos[1] %in% 1:4)
+				mar[key.pos[1]] = mar[key.pos[1]] + .5
 			par(mar = mar)
 
 			# plot the map:
@@ -175,12 +208,14 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 			if (! draw.key)
 				key.pos = NULL
 			lt = sf::.get_layout(st_bbox(x), dims[3], par("din"),
-						if (join_zlim && key.pos.missing) -1 else key.pos, key.width, mfrow = mfrow)
+						if (join_zlim && key.pos.missing) -1 else key.pos[1], key.width, mfrow = mfrow, main = main)
+			if (key.pos.missing)
+				key.pos = lt$key.pos
 			title_size = if (is.null(main))
 					0
 				else
 					1.2
-			layout(lt$m, widths = lt$widths, heights = lt$heights, respect = FALSE)
+			layout(lt$m, widths = lt$widths, heights = lt$heights, respect = compact)
 			par(mar = c(axes * 2.1, axes * 2.1, title_size, 0))
 			labels = st_get_dimension_values(x, 3, center = center_time)
 			for (i in seq_len(dims[3])) {
@@ -212,19 +247,21 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 							 bbox = st_bbox(x))
 					}
 				}
-				box(col = box_col)
+				if (!is.na(box_col))
+					box(col = box_col)
 			}
 			for (i in seq_len(prod(lt$mfrow) - dims[3])) # empty panels:
 				plot.new()
+
 			if (draw.key) {
 				if (missing(key.lab) && inherits(x[[1]], "units"))
 					key.lab = units::make_unit_label(names(x)[1], x[[1]])
 				values = structure(x[[1]], dim = NULL)
 				if (is.factor(values))
-					.image_scale_factor(levels(values), col, key.pos = lt$key.pos,
+					.image_scale_factor(levels(values), col, key.pos = key.pos,
 						key.width = key.width, key.length = key.length, axes = axes,...)
 				else
-					.image_scale(values, col, breaks = breaks, key.pos = lt$key.pos,
+					.image_scale(values, col, breaks = breaks, key.pos = key.pos,
 						key.width = key.width, key.length = key.length, axes = axes,..., lab = key.lab)
 			}
 		}
@@ -237,11 +274,11 @@ plot.stars = function(x, y, ..., join_zlim = TRUE, main = make_label(x, 1), axes
 			key.width = key.width, reset = reset, axes = axes, main = main)
 	} else
 		stop("no raster, no features geometries: no default plot method set up yet!")
-	if (reset) {
+	if (reset && need_to_reset) {
 		layout(matrix(1)) # reset
-		desel = which(names(opar) %in% c("cin", "cra", "csi", "cxy", "din", "page", "pin"))
-		par(opar[-desel])
+		par(opar)
 	}
+	invisible(NULL)
 }
 
 get_breaks = function(x, breaks, nbreaks, logz = NULL) {
@@ -298,7 +335,7 @@ image.stars = function(x, ..., band = 1, attr = 1, asp = NULL, rgb = NULL,
 		xlim = st_bbox(extent)$xlim, ylim = st_bbox(extent)$ylim, text_values = FALSE,
 		text_color = 'black', axes = FALSE,
 		interpolate = FALSE, as_points = FALSE, key.pos = NULL, logz = FALSE,
-		key.width = lcm(1.8), key.length = 0.618, add.geom = NULL, border = NA,
+		key.width = kw_dflt(x, key.pos), key.length = 0.618, add.geom = NULL, border = NA,
 		useRaster = isTRUE(dev.capabilities()$rasterImage == "yes"), extent = x) {
 
 	dots = list(...)
@@ -461,6 +498,8 @@ image.stars = function(x, ..., band = 1, attr = 1, asp = NULL, rgb = NULL,
 			col = text_color)
 	}
 	if (axes) { # FIXME: see sf::plot.sf for refinements to be ported here?
+		if (!is.null(dots$cex.axis))
+			par(cex.axis = dots$cex.axis)
         if (isTRUE(st_is_longlat(x))) {
 			if (isTRUE(all.equal(st_bbox(x), st_bbox(), tolerance = .1, 
 								 check.attributes = FALSE))) {