+++ R/rasterize.R
@@ -20,14 +20,15 @@ st_align = function(bb, d) {
 #' rasterize simple feature geometries
 #' @export
 #' @param sf object of class \code{sf}
-#' @param template stars object with desired target geometry, or target geometry alignment if \code{align=TRUE}
+#' @param template optional; stars object with desired target geometry, or target geometry alignment if \code{align=TRUE}; see details
 #' @param file temporary file name
 #' @param driver driver for temporary file
 #' @param options character; options vector for \code{GDALRasterize}
-#' @param align logical; if \code{TRUE}, \code{template} contain the geometry alignment, 
-#' informing target resolution and offset only.
+#' @param align logical; if \code{TRUE}, \code{template} is only used for the geometry _alignment_, 
+#' informing target resolution and offset 
 #' @param proxy logical; should a proxy object be returned?
 #' @param ... arguments passed on to \link{st_as_stars}
+#' @details if `template` is a `stars` object, non-NA cells that are not covered by `sf` receive the value in `template`; see also argument `align`.
 #' @examples
 #' demo(nc, echo = FALSE, ask = FALSE)
 #' (x = st_rasterize(nc)) # default grid:
@@ -62,9 +63,9 @@ st_rasterize = function(sf, template = guess_raster(sf, ...) %||%
 		bb = st_align(st_bbox(sf), d <- st_dimensions(st_normalize(template)))
 		dx = d[[1]]$delta
 		dy = d[[2]]$delta
-		template = st_as_stars(bb, values = 0.0, dx = dx, dy = dy,
-			nx = round(diff(bb[c("xmin", "xmax")])/dx), 
-			ny = round(diff(bb[c("ymin", "ymax")])/dy),
+		template = st_as_stars(bb, values = NA, dx = dx, dy = dy,
+			nx = round(diff(bb[c("xmin", "xmax")])/abs(dx)), 
+			ny = round(diff(bb[c("ymin", "ymax")])/abs(dy)),
 			proxy = proxy)
 	} else
 		template = st_normalize(template)
@@ -82,7 +83,7 @@ st_rasterize = function(sf, template = guess_raster(sf, ...) %||%
 	attrs = as.data.frame(sf)[-geoms]
 	for (i in which(sapply(sf, inherits, "factor"))) # factors:
 		sf[[i]] = as.numeric(sf[[i]])
-	sf::gdal_rasterize(sf, template, get_geotransform(template), file, driver, options)
+	sf::gdal_rasterize(sf, template, st_geotransform(template), file, driver, options)
 	ret = read_stars(file, driver = driver, proxy = proxy)
 	if (!proxy) {
 		if (length(dim(ret)) > 2)
@@ -135,75 +136,111 @@ guess_raster = function(x, ...) {
 #' @export
 #' @param dims the column names or indices that form the cube dimensions
 #' @param coords same as dims, for symmetry with \link[sf]{st_as_sf}
-#' @param xy the x and y raster dimension names or indices; only takes effect after dims has been specified
+#' @param xy the x and y raster dimension names or indices; only takes effect after \code{dims} has been specified, see details
 #' @param y_decreasing logical; if TRUE, (numeric) y values get a negative delta (decrease with increasing index)
 #' @name st_as_stars
+#' @details
+#' If \code{xy} is not specified and the first two dimensions in \code{dims} are both numeric,
+#' then it is set to these two dimensions.
 #' @examples
 #' if (require(plm, quietly = TRUE)) {
-#'   data(Produc, package = "plm")
-#'   st_as_stars(Produc, y_decreasing = FALSE)
 #'  data(Produc, package = "plm")
-#'  st_as_stars(Produc, y_decreasing = FALSE)
+#'  st_as_stars(Produc)
 #' }
-st_as_stars.data.frame = function(.x, ..., dims = coords, xy = dims[1:2], y_decreasing = TRUE, coords = 1:2) {
+#' if (require(dplyr, quietly = TRUE)) {
+#'   # https://stackoverflow.com/questions/77368957/
+#' spatial_dim <- st_sf(
+#'   ID = 1:3,
+#'   geometry = list(
+#'     st_polygon(list(
+#'       cbind(c(0, 1, 1, 0, 0), c(0, 0, 1, 1, 0))
+#'     )),
+#'     st_polygon(list(
+#'       cbind(c(1, 2, 2, 1, 1), c(0, 0, 1, 1, 0))
+#'     )),
+#'     st_polygon(list(
+#'       cbind(c(2, 3, 3, 2, 2), c(0, 0, 1, 1, 0))
+#'     ))
+#'   )
+#' )
+#' weekdays_dim <- data.frame(weekdays = c("Monday", "Tuesday", "Wednesday", 
+#'     "Thursday", "Friday", "Saturday", "Sunday"))
+#' hours_dim <- data.frame(hours = c("8am", "11am", "4pm", "11pm"))
+#' sf_dta <- spatial_dim |>
+#'   cross_join(weekdays_dim)|>
+#'   cross_join(hours_dim) |>
+#'   mutate(population = rnorm(n(), mean = 1000, sd = 200)) |>
+#'   select(everything(), geometry)
+#' 
+#' st_as_stars(sf_dta, dims = c("weekdays", "hours", "geometry"))
+#' }
+#' demo(nc, echo=FALSE,ask=FALSE)
+#' st_as_stars(nc)
+#' st_as_stars(st_drop_geometry(nc), dims = "NAME")
+#' data.frame(expand.grid(x=1:5, y = 1:5), z = rnorm(25)) |> st_as_stars()
+st_as_stars.data.frame = function(.x, ..., dims = coords, xy, y_decreasing = TRUE, coords = 1:2) {
 
 	if (missing(dims) && !missing(xy))
 		stop("parameter xy only takes effect when the cube dimensions are set with dims")
 	if (is.character(dims))
 		dims = match(dims, names(.x))
-	if (is.character(xy))
+	if (missing(xy)) {
+		if (length(dims) > 1 && all(sapply(.x[dims[1:2]], is.numeric)))
+			xy = dims[1:2]
+		else 
+			xy = c(-1, -1) # cancels y_decreasing
+	}
+	if (is.character(xy)) {
 		xy = match(xy, names(.x))
-	if (any(is.na(xy)))
-		stop("xy coordinates not found in data")
+		if (any(is.na(xy)))
+			stop("xy coordinates not found in data")
+	}
+	stopifnot(length(dims) >= 1, all(dims >= 1), !any(is.na(dims)))
 
 	index = NULL
 	dimensions = list()
-	if (length(dims) >= 2) {
-		this_dim = 1
-		for (i in dims) {
-			v = .x[[i]]
-			if (inherits(v, "sfc")) {
-    			if (!requireNamespace("digest", quietly = TRUE))
-        			stop("package digest required, please install it first") # nocov
-				dig = sapply(v, digest::digest)
-				uv = unique(dig) # no need to sort
-				ix = match(dig, uv) # but look up "hash collision"
-			} else {
-				suv = sort(unique(v), decreasing = y_decreasing && i == xy[2])
-				ix = match(v, suv)
-			}
-			index = cbind(index, ix)
-			dimensions[[this_dim]] = if (inherits(v, "sfc"))
-					create_dimension(values = v[match(uv, dig)])
-				else
-					create_dimension(values = suv, is_raster = i %in% xy)
-			this_dim = this_dim + 1
+	this_dim = 1
+	for (i in dims) {
+		v = .x[[i]]
+		if (inherits(v, "sfc")) {
+    		if (!requireNamespace("digest", quietly = TRUE))
+       			stop("package digest required, please install it first") # nocov
+			dig = sapply(st_equals(v), digest::digest)
+			uv = unique(dig) # no need to sort
+			ix = match(dig, uv) # but look up "hash collision"
+		} else {
+			suv = if (is.factor(v))
+					levels(v)
+				else if (is.character(v))
+					unique(v)
+				else # numeric:
+					sort(unique(v), decreasing = y_decreasing && i == xy[2])
+			ix = match(v, suv)
 		}
-		names(dimensions) = names(.x)[dims]
-	
-		raster_xy = if (length(xy) == 2) names(.x)[xy] else c(NA_character_, NA_character_)
-		d = create_dimensions(dimensions, raster = get_raster(dimensions = raster_xy))
-		l = lapply(.x[-dims], function(x) {
-				m = if (is.factor(x))
-						structure(factor(rep(NA_character_, prod(dim(d))), levels = levels(x)),
-							dim = dim(d))
-					else 
-						array(NA, dim = dim(d))
-				m[index] = x # match order
-				m 
-			}
-		)
-	} else {
-		l = lapply(.x, as.array)
-		dimensions[[1]] = if (length(dims) == 0 || dims < 1)
-				create_dimension(values = row.names(.x))
-			else {
-				l[[dims]] = NULL
-				create_dimension(values = .x[[dims]])
-			}
-		names(dimensions) = "rows"
-		d = create_dimensions(dimensions)
+		index = cbind(index, ix)
+		dimensions[[this_dim]] = if (inherits(v, "sfc"))
+				create_dimension(values = v[match(uv, dig)])
+			else
+				create_dimension(values = suv, is_raster = i %in% xy)
+		this_dim = this_dim + 1
 	}
+	names(dimensions) = names(.x)[dims]
+	
+	raster_xy = if (length(xy) == 2 && all(xy > 0)) 
+			names(.x)[xy] 
+		else 
+			c(NA_character_, NA_character_)
+	d = create_dimensions(dimensions, raster = get_raster(dimensions = raster_xy))
+	l = lapply(.x[-dims], function(x) {
+			m = if (is.factor(x))
+					structure(factor(rep(NA_character_, prod(dim(d))), levels = levels(x)),
+						dim = dim(d))
+				else 
+					array(NA, dim = dim(d))
+			m[index] = x # match order
+			m 
+		}
+	)
 	st_stars(l, d)
 }