+++ R/subset.R
@@ -176,7 +176,8 @@
 		stop("selector i should be a stars object or a lenght-one integer or character vector")
 }
 
-st_intersects.bbox = function(x, y, ...) {
+#' @export
+st_intersects.bbox = function(x, y, ...) { # FIXME: segmentize first if geographic coords? see sf::st_sample
 	if (!inherits(y, "sfc"))
 		y = st_as_sfc(y)
 	st_intersects(st_as_sfc(x), y, ...)
@@ -295,7 +296,7 @@ st_crop.stars = function(x, y, ..., crop = TRUE, epsilon = sqrt(.Machine$double.
 	if (!inherits(y, "bbox")) { # post-process: burn in geometry mask
 		dxy = attr(dm, "raster")$dimensions
 		xy_grd = if (is_curvilinear(x) || !as_points) # FIXME: for curvilinear as_points should work too!
-				st_as_sfc(st_dimensions(x)[dxy], as_points = as_points, geotransform = get_geotransform(x))
+				st_as_sfc(st_dimensions(x)[dxy], as_points = as_points, geotransform = st_geotransform(x))
 			else
 				st_as_sf(do.call(expand.grid, expand_dimensions.stars(x)[dxy]), coords = dxy, crs = st_crs(x))
 		inside = st_intersects(st_union(y), xy_grd)[[1]]